rename: payment 상수 파일 이름 변경
